add to e2e test (#360)
